# [RyuJIT] Make UNBOX_CALL block cold in case of inlined check