--- a/src/coreclr/src/jit/importer.cpp
+++ b/src/coreclr/src/jit/importer.cpp
@@ -15601,7 +15601,7 @@ void Compiler::impImportBlockCode(BasicBlock* block)
                     }
                     op1 = gtNewHelperCallNode(helper, TYP_VOID, gtNewCallArgs(op2, op1));
 
-                    op1 = new (this, GT_COLON) GenTreeColon(TYP_VOID, gtNewNothingNode(), op1);
+                    op1 = new (this, GT_COLON) GenTreeColon(TYP_VOID, gtNewNothingNode(), op1, true);
                     op1 = gtNewQmarkNode(TYP_VOID, condBox, op1);
 
                     // QMARK nodes cannot reside on the evaluation stack. Because there