@AndyAyersMS 2020-10-01
I like the idea here, but have some suggestions.

The helper does more than throw, it handles cases where types are considered to be equivalent (one example is an enum and its underlying type). So you may want to try benchmarking a case like this too, and see what happens to the perf.

On the implementation side, I'd prefer that we add a `likelihood` field to the `GT_COLON` which is always set -- say to estimate how often the predicate is true. It can default to 0.50 (50 in our current scheme). Then use that value over in `fgExpandQmarkStmt` instead of the hard-coded numbers we have there now, and fix any other `GT_COLON` cases that want to bias probabilities to work in this scheme.

(There may be some extra work involved if we ever end up inverting the sense of the `GT_COLON` predicate; the likelihood will also need to be inverted. I suspect we don't do this, but can't say for sure).

For this case, I would like to see us use a small but nonzero likelihood, but if you do that you'll see other issues. See what happens if you use say 95/5 bias and the unbox is in a loop:
```C#
int result = 0;
for (int i = 0; i < N; i++)
{
    result += (int) o;
}
return result;
```

Along those likes I plan to change `inheritWeightPercentage` to set the `BBF_PROF_WEIGHT` on the target if it was set on the source. I'll probably PR that separately, fairly soon.