@EgorBo 2020-10-01
@AndyAyersMS ah, I totally forgot about Enums, here is the new benchmark:
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestValues))]
public int UnboxAsInt(object o1, object o2, object o3)
{
    return (int)o1 + (int)o2 + (int)o3;
}

[Benchmark]
[ArgumentsSource(nameof(TestValues))]
public MethodImplOptions UnboxAsEnum(object o1, object o2, object o3)
{
    return (MethodImplOptions)o1 | (MethodImplOptions)o2 | (MethodImplOptions)o3;
}

public static IEnumerable<object[]> TestValues()
{
    yield return new object[] { MethodImplOptions.AggressiveInlining, MethodImplOptions.NoInlining, MethodImplOptions.InternalCall };
    yield return new object[] { MethodImplOptions.AggressiveInlining, 8, 4096 };
    yield return new object[] { 256, 8, 4096 };
}
```
Current results (master)
```
|      Method |                 o1 |         o2 |           o3 |       Mean |
|------------ |------------------- |----------- |------------- |-----------:|
|  UnboxAsInt |                256 |          8 |         4096 |  0.8444 ns | // inlined check
|  UnboxAsInt | AggressiveInlining |          8 |         4096 |  4.2484 ns | // uses CORINFO_UNBOX
|  UnboxAsInt | AggressiveInlining | NoInlining | InternalCall | 11.4077 ns | // uses CORINFO_UNBOX

| UnboxAsEnum |                256 |          8 |         4096 | 12.5139 ns | // uses CORINFO_UNBOX
| UnboxAsEnum | AggressiveInlining |          8 |         4096 |  8.9449 ns | // uses CORINFO_UNBOX
| UnboxAsEnum | AggressiveInlining | NoInlining | InternalCall |  1.1106 ns | // inlined check
```
So this PR actually makes `enum->box->int` or `int->box->enum` operations noticeably slower.
I just wonder if we can add a second "inlined check" for enum, something like:
```
if (obj is int)
    objAsInt
else if (obj is Enum)
    objAsEnum
else
   CORINFO_UNBOX
```